+++ src/glue/auth.cpp
@@ -28,6 +28,10 @@ auth::Permission PrivilegeToPermission(query::AuthQuery::Privilege privilege) {
       return auth::Permission::REPLICATION;
     case query::AuthQuery::Privilege::LOCK_PATH:
       return auth::Permission::LOCK_PATH;
+    case query::AuthQuery::Privilege::READ_FILE:
+      return auth::Permission::READ_FILE;
+    case query::AuthQuery::Privilege::FREE_MEMORY:
+      return auth::Permission::FREE_MEMORY;
     case query::AuthQuery::Privilege::AUTH:
       return auth::Permission::AUTH;
   }